Splitting share group validator in two separate validators
Presets should not yet be validated for the fact if they're explictly setting all the shares.

Discussed here: https://github.com/quintel/atlas/pull/86\#issuecomment-275112404
