+++ lib/atlas.rb
@@ -42,12 +42,14 @@ require_relative 'atlas/active_document/subclassing'
 require_relative 'atlas/active_document/manager'
 require_relative 'atlas/active_document/production_manager'
 require_relative 'atlas/active_document/query_validator'
+require_relative 'atlas/active_document/share_group_validator'
+require_relative 'atlas/active_document/share_group_inclusion_validator'
+require_relative 'atlas/active_document/share_group_total_validator'
 require_relative 'atlas/active_document'
 require_relative 'atlas/collection'
 require_relative 'atlas/value_object'
 
 require_relative 'atlas/edge'
-require_relative 'atlas/share_group_validator'
 require_relative 'atlas/input_helper'
 require_relative 'atlas/input'
 require_relative 'atlas/initializer_input'
